requirements/base.txt
altgraph==0.16.1
boto3==1.9.71
botocore==1.12.71
Django==2.1.4
django-appconf==1.0.2
django-compressor==2.2
djangorestframework==3.9.0
docutils==0.14
future==0.17.1
jmespath==0.9.3
macholib==1.11
pefile==2018.8.8
Pillow==5.3.0
psycopg2==2.7.6.1
PyAutoGUI==0.9.39
PyInstaller==3.4
PyMsgBox==1.0.6
PyScreeze==0.1.18
python-dateutil==2.7.5
python-decouple==3.1
PyTweening==1.0.3
pytz==2018.7
pywin32-ctypes==0.2.0
rcssmin==1.0.6
rjsmin==1.0.12
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
win32core==221.36
win32gui==221.6
